changed the first two jsx files into tsx
